Fix logic for chaining calculations
Before this change, in a scenario where the user has already inputted their first operand, but accidentally pressed an unintended operator for their calculation, attempting to change the operator to another one by selecting a new operator BEFORE inputting their second operand would result in erroneous calculations. This was because the conditional logic for calculating chained operations and displaying the answer in real-time on the display was setting the second operand to be 0 regardless of whether a calculation had even been made in the first place. With this change, the user should be able to change their intended operator choice midway through the calculation.