Add feature for equals sign to increment
On most modern calculators, when the user presses the equals sign repeatedly after making a calculation, the answer will increment by the last operand and operator that the user inputted. Before this change, pressing the equals sign repeatedly did nothing. With this change, the equals sign will increment with the last operand and operator it remembers, in line with modern calculators.