Fix dividing decrement feature in equals sign
Before this change, when the user divide continuously by repeatedly pressing the division button after performing an operation with two operands and the divide operator, the display would correctly decrement the result using the last inputted operand and the divide operator in memory. However, if the user inputs divide or multiply after the last equals sign input, the memory and display would be cleared to all 0s. This change fixed the issue by adding more conditions so the calculator now properly increments and decrements with the equals sign, while also allowing the user to go back to regular operations afterwards.