Reorganize code structure and add comments
These changes reorganized a large portion of the code logic including creating new helper functions to minimize code repetition.

All of the previous conditional logic was based on whether the operands and values were truthy or falsy. The problem with this was that it did not accept the number 0 as an operand because it is falsy. This was fixed by including 0 as a valid value in the new exists function. This change alone created a cascading effect which affected other portions of the code, requiring changes in multiple places for the calculator to be in working condition again.

The equals pressed and divide logic was also removed from the operation function, along with many associated nested conditions to clean up the code structure. This makes the code more readable while also minimizing redudancy. The compute function also received similar treatment. Some features are temporarily removed as a result of this change, but they will be integrated back using the new logic described above.

Finally, minor changes were done to clear, delete, and invert sign functions along with added comments to many parts of the code for clarity.