Reintegrate equals pressed logic to allow for chaining operations
This change reintegrated the equals pressed switch logic back into the input and operator functions.

Equals pressed is switched on only when the equals button is explicity pressed by the user for their computation. This is needed because the number input and operator functions need to do different things depending on if the user is just starting their operation, or if they are chaining their operation after a previously finished computation.

The equals pressed check in the number input function is to ensure that any number button input following an operation is invalid since, realistically, anything following pressing the equals button should expect either an operator to chain operations using the result of the computation as an operand, or the end of the operation. A number input is invalid and will for now be regarded as starting anew by clearing memory.

The equals pressed check in the operator function is for chaning operations. There are two ways to compute and display the answer on the calculator: explicitly pressing equals or chaining operations without pressing equals. By pressing equals, the operation is only performed if all operands are valid and an operator exists, which will compute the operation and display it. This also has the added feature of saving the last/second inputted operand and operator in memory, allowing for repeated increment/decrement calculations using only the equals button. Chaining operations without pressing equals still calls the compute function to compute the operation and display the answer, but it also sets up the answer to that computation as the first operand for further chained operations.