Reintegrate end of calc, previously equals pressed, into functions
End of calc serves to keep track of whether the calculation has ended, changing what the functions do depending on the situation. The calculation is considered ended when the user presses the equals sign.

This change added the end of calc check in the number input button to allow the user to immediately start a new operation with new operands if they just finish their previous computation. The end of calc check was also added at the beginning of the the operation function for the same purpose. If a user inputs a number input after pressing equals, their intention is most likely to start a new operation. Inputting an operator following their number input clearly shows their intention to begin a new operation, therefore, the end of calc checks will disregard the the results from the previous computation.

Finally, a minor feature was added in the compute function for delegating operands in order to mimic what other calculators can do. With this change, if the user inputs a number for their first operand, inputs an operator, then the equals button, the first operand would also be set as the second operand and the operation will be performed.

Overall, these changes serve to mimic functionality that most common calculators have, such as the iPhone and Windows calculators.