Fix invert sign function to allow for resetting operations with decimals
Prior to this change, if the user inputs a decimal after their operation ends, the calculator considers it as starting a new operation, which is the intended effect. However, a problem arose if the user inverts the sign of their decimal point number input, resulting in unintended loss of numbers.

This was because the invert sign function, in the case where it is the end of calculation, only dealt with inverting the first operand and using that as an operand for chaining calculations. Meanwhile, the add decimal function used current input value to concatenate the decimal point. This change allow users to input negative floating numbers after finishing their operation by reconciling the firstOperand/currentInputValue delegation in the invert sign, decimal and number input functions for the case of when the calculation has ended, and the user intends to start a new operation.