Fix addition and subtraction for cases where second operand is 0
Prior to this change, the equals button would still increment/decrement even if the second operand is 0 because one of the conditions for assigning second operand was only looking at whether the second operand was falsy for assignment. This change fixed the condition to check with the exists function instead of just falsy, to accepts the number 0 as a valid user inputted operand.