Fix feature that allow user to continue operations at the end
This calculator has the feature where, after the user performs an operation, they could use the answer from that computation as the first operand for a new operation by pressing the operator button. Alternatively, the user could also input a new number, which, for our purpose, is considered as starting a brand new operation the moment they follow the number input with an operator input. Prior to this change, the calculator was only saving the answer from the last computation and using that as the first operand, regardless of whether the user inputs a new number or not, even if their intention is to start a new operation. This change fixed that feature and allows for both, whilst also taking into account whether the user inverts the sign of their input or not.